+++ src/interfaces/nano-data-binding.ts
@@ -64,3 +64,5 @@ export interface Changes {
     added: any[],
     removed: any[]
 }
+
+export type MockArr = {[key: string]: number}[]