+++ CHANGES.md
@@ -1,10 +1,15 @@
 # 0.0.9
 * Better syntax - Origin prefixes in the attributes. It was confusing to read the origin tokens [".", ":", "$"] especially for the short notations. Attributes prefixes are far easier to read ["p-if", "e-if", "o-if"]. Side effect of this change, simpler code, easier to maintain.
-* Restore optional debug logs.
+* Restore debug logs.
 * Split in smaller files.
-* Automatically init data binds. Smart defaults over configuration. Eliminates the need to have unique class or id identifiers. Exceptions can be handled manually.
-    * Observe DOM mutations, bind the new elements.
+* Automatically init data binds. 
+    * Smart defaults over configuration. Eliminates the need to have unique class or id identifiers. Exceptions can be handled manually.
+    * Observe DOM mutations, bind the new elements. Triggered by nano data binds.
     * Search for the closest web component
+    * Instead of copying the references, methods are invoked with binding the children context.
+    * To get rid of copying references we need to replace all inline event handlers with custom ones. `onclick=""` becomes `(click)=""`
+    * Data bind throws error if invoked method is not defined on parent context.
+    * `_nano_no-auto-init` flag can be used to disable the autobind behavior for automatic testing purposes.
 * Data binding to context properties.
 * Default to context property if no origin is defined in the syntax.
 * Adding examples